cv_agents/pyproject.toml
[project]
name = "cv_agents"
version = "0.1.0"
description = "cv_agents using crewAI"
authors = [{ name = "Your Name", email = "you@example.com" }]
requires-python = ">=3.10,<3.13"
dependencies = [
    "crewai[tools]>=0.108.0,<1.0.0",
    "PyGithub>=2.1.1",
    "together>=0.2.0",
    "pydantic>=2.0.0",
    "python-dotenv>=1.0.0",
    "PyYAML>=6.0.1",
   "GitPython>=3.1.40",

    # LangChain tools
    "langchain>=0.1.0",
    "langchain-community>=0.0.10",
    "faiss-cpu>=1.7.4",  # Pour la recherche vectorielle
    
    # Utilities
    "requests>=2.31.0",
    "beautifulsoup4>=4.12.2",  # Pour le web scraping
    "markdown>=3.5.1",
]


[project.scripts]
cv_agents = "cv_agents.main:run"
run_crew = "cv_agents.main:run"
train = "cv_agents.main:train"
replay = "cv_agents.main:replay"
test = "cv_agents.main:test"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.crewai]
type = "crew"
